# packages/contracts/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Prepeet API
  version: "1.0.0"
  description: |
    The browser talks to this API and to nothing else. It never reaches
    PostgreSQL, Temporal, object storage or the intelligence service directly,
    which is what keeps authorization decidable in one place.

    Every request operates under exactly one explicit active tenant. Tenant is
    never inferred from a resource identifier.
  contact:
    name: Prepeet engineering
servers:
  - url: /api/v1
    description: Version in the path. A new version is a project rather than a release.
...
paths:
...
  /auth/refresh:
    post:
      tags: [authentication]
      operationId: refresh
      summary: Rotate the session using a refresh token
      description: |
        Each refresh issues a new pair and retires the old one. Presenting a
        retired token means either a stolen token or a client bug, and both are
        handled the same way: the whole session family is revoked and the event
        is audited. Being logged out is a cheap failure; an attacker keeping a
        foothold is not.
      security: []
      responses:
        "200":
          description: A new session was issued and the old one retired.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            Set-Cookie:
              description: Rotated session and refresh cookies.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: |
            The refresh token was missing, expired, or already used. A reused
            token also revokes the entire family.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: prepeet_session
      description: |
        An opaque session token, stored hashed server side so a database read
        yields nothing presentable. Opaque rather than a JWT because membership
        revocation must take effect within seconds. See ADR-0003.
...
  headers:
    CacheControl:
      description: |
        What an intermediary may do with this response, per the caching
        conventions in ADR-0004.

        Declared on every response rather than left to the server, because a
        cacheability that exists only in a handler is a property nobody can
        review and no client can rely on. What the server actually sends is
        asserted against this document, so the two cannot drift.
      required: true
      schema:
        type: string
...
  schemas:
    ErrorCode:
      type: string
      description: |
        Stable and never reused for a different meaning. Generated into both
        clients as an enum, so a client can handle codes exhaustively and a new
        one is a compile-time prompt rather than a runtime surprise.

        Messages are for humans and may be reworded at any time. Only the code
        is machine logic.
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - METHOD_NOT_ALLOWED
        - IDEMPOTENCY_CONFLICT
        - RATE_LIMITED
        - CREDENTIALS_INVALID
        - SESSION_EXPIRED
        - REFRESH_TOKEN_REUSED
        - TOKEN_INVALID
        - TOKEN_EXPIRED
        - TOKEN_USED
        - TOKEN_SUPERSEDED
        - CODE_INCORRECT
        - CODE_ATTEMPTS_EXHAUSTED
        - RESEND_COOLING_DOWN
        - INTERNAL
    FieldError:
      type: object
      description: One field-level validation failure.
      required: [field, code, message]
      properties:
        field:
          type: string
          description: JSON pointer to the offending field.
          examples: ["/email"]
        code:
          type: string
          examples: ["REQUIRED"]
        message:
          type: string
    Error:
      type: object
      description: The single error envelope every failure in this API uses.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable, field_errors, request_id]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Human readable, never machine logic.
            retryable:
              type: boolean
              description: Whether retrying the same request could plausibly succeed.
            field_errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
            request_id:
              type: string
              description: |
                Correlation identifier, also returned in X-Request-ID. It is
                what makes a user's support report resolvable to one trace.
              examples: ["req_01a0301daa1070009"]
...
    Session:
      type: object
      description: |
        Describes the session that was established. The tokens themselves are
        in HTTP-only cookies and never appear in a response body, so no script
        in the page can read them.
      required: [user_id, expires_at, authenticated_at]
      properties:
        user_id:
          type: string
          format: uuid
        active_tenant_id:
          type: [string, "null"]
          format: uuid
          description: |
            Null for a candidate who belongs to no tenant. Every request
            operates under exactly one explicit active tenant, and it is never
            inferred from a resource identifier.
        expires_at:
          type: string
          format: date-time
        authenticated_at:
          type: string
          format: date-time
          description: |
            When the user last proved who they are, which is not when the
            session was issued. Destructive and evaluation-changing capabilities
            require this to be recent.
...
security:
  - sessionCookie: []